Verander die uitdruk beheerstok

+++ beheerRaspberryPi/joyStuur.py
@@ -75,7 +75,7 @@ while done==False:
     clock.tick(20)
     axis = joystick.get_axis( 1 )
     # Skakel die beheerstok as oor na 'n servo hoek
-    servo_0_angle = axis*90 + 90
+    servo_0_angle = int(axis*90 + 90)
     print("Axis {} value: {:>6.2f}".format(1, servo_0_angle))
